Oral-B: Add support for "IO Series 7"
This change introduces a new way to map the "Modes" values based on the device model.

Support for the SmartSeries 7000 has been retained and support for the IO Series 7 has been added. It is likely that some mappings are missing for IO Series 8 & 9, however I do not have these devices on-hand.

The value for the model should really go into the dveice_type field, however this caused the whole plug-in to break for me (attributes stopped reporting). I instead placed the model into the "Firmware" field which worked and would've otherwise just been "Oral-B". The second issue with this is that we do not know the model until a broadcast is received.

Open to suggestions on ways to improve this commit as I am not a Home Assistant, BLE or Python expert by any stretch.